+++ GameEngine/src/gamelogic/SpaceObject.java
@@ -21,10 +21,10 @@ public abstract class SpaceObject {
     public final String packageName;
     public final String className;
     
-    public int energy; // Energy that aliens gain every turn from the planet
-    public int tech; // Tech boost for the planet
+    public double energy; // Energy that aliens gain every turn from the planet
+    public double tech; // Tech boost for the planet
     
-    public SpaceObject (GameVisualizer vis, int x, int y, String packageName, String className, int energy, int tech) {
+    public SpaceObject (GameVisualizer vis, int x, int y, String packageName, String className, double energy, double tech) {
         this.x = x;
         this.y = y;
         this.energy = energy;
